changed loop

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -9,7 +9,7 @@ void rev_string(char *s)
 {
 	int i;
 	int j;
-	float length;
+	int length;
 	char nrml;
 	char revrs;
